BloomFilter.computeBFs and .computeNodeBF simplified
Since all operations on the byte[] Bloom filters transform them into
BigIntegers before changing anything, there is no need to specify the
exact size of the empty initial byte[] to avoid outOfBoundsExceptions.
Therefore the s argument of computeBFs and size argument of
computeNodeBF are not needed and have been removed.